my-ocr-project/requirements.txt
Flask
pytesseract
pdfminer.six
Pillow